requirements.txt
numpy
pandas
python-box
pyyaml
tensorflow
mkdocs
seaborn
matplotlib
scipy
coverage
scikit-learn
json2html